requirements/dev.txt
pyserial
crcmod
grpcio
grpcio-tools
jsonschema
dateutils
bitarray
loguru
netifaces
asciimatics
2to3
